P1/requirements.txt
fastapi
uvicorn #web server that will run the FastAPI app
ffmpeg-python
numpy
pillow
scipy
pywavelets
python-multipart
